+++ CHANGELOG.md
@@ -226,3 +226,4 @@
 * Added method to get list of supported tokens
 * Added method to get exchange rates
 * Added method to get slippage
+* Added method to get estimated gas